+++ orttraining/orttraining/test/python/_test_commons.py
@@ -2,6 +2,10 @@ import math
 import os
 import sys
 import subprocess
+import copy
+import numpy as np
+
+from onnxruntime.training import optim
 
 def is_windows():
     return sys.platform.startswith("win")
@@ -76,3 +80,57 @@ def legacy_poly_lr_scheduler(global_step, initial_lr, total_steps, warmup, power
         decay = lr_range * pct_remaining ** power + lr_end
         new_lr = decay
     return new_lr
+
+
+def generate_dummy_optim_state(model, optimizer):
+    if not (isinstance(optimizer, optim.AdamConfig) or isinstance(optimizer, optim.LambConfig)):
+        return dict()
+
+    moment_keys = ["Moment_1", "Moment_2"]
+    uc_key = "Update_Count"
+    step_key = "Step"
+    shared_state_key = "shared_optimizer_state"
+
+    optim_state = dict()
+    name_to_initializer_map = {n.name:n for n in model.graph.initializer}
+    initializers_names = name_to_initializer_map.keys()
+    for weight in initializers_names:
+        per_weight_state = dict()
+        weight_shape = name_to_initializer_map[weight].dims
+        for moment in moment_keys:
+            per_weight_state[moment] = np.full(weight_shape, 2.5, dtype=np.float32)
+        if isinstance(optimizer, optim.AdamConfig):
+            per_weight_state[uc_key] = np.full([1], 5, dtype=np.int64)
+        optim_state[weight] = copy.deepcopy(per_weight_state)
+    if isinstance(optimizer, optim.LambConfig):
+        step_val = np.full([1], 5, dtype=np.int64)
+        optim_state[shared_state_key] = {step_key : step_val}
+    return optim_state
+
+
+def get_optim_state_from_state_dict(state_dict, optimizer):
+    if not (isinstance(optimizer, optim.AdamConfig) or isinstance(optimizer, optim.LambConfig)):
+        return dict()
+    
+    moment_keys = ["Moment_1", "Moment_2"]
+    uc_key = "Update_Count"
+    step_key = "Step"
+    shared_state_key = "shared_optimizer_state"
+
+    optim_state = dict()
+    for param_name, v in state_dict.items():
+        for moment in moment_keys:
+            if param_name.startswith(moment):
+                fp32_name = param_name.split(moment + '_')[-1]
+                if fp32_name not in optim_state:
+                    optim_state[fp32_name] = dict()
+                optim_state[fp32_name].update({moment: v})
+                break
+        if param_name.startswith(uc_key):
+            fp32_name = param_name.split(uc_key + '_')[-1]
+            if fp32_name not in optim_state:
+                optim_state[fp32_name] = dict()
+            optim_state[fp32_name].update({uc_key: v})
+        elif param_name == step_key:
+            optim_state[shared_state_key] = {step_key: v}
+    return optim_state
