+++ orttraining/orttraining/test/python/orttraining_test_orttrainer_bert_toy_onnx.py
@@ -45,58 +45,6 @@ def generate_random_input_from_model_desc(desc, seed=1, device = "cuda:0"):
         sample_input.append(torch.randint(0, num_classes[index], tuple(size), dtype=dtype).to(device))
     return sample_input
 
-def generate_dummy_optim_state(model, optimizer):
-    if not (isinstance(optimizer, optim.AdamConfig) or isinstance(optimizer, optim.LambConfig)):
-        return dict()
-
-    moment_keys = ["Moment_1", "Moment_2"]
-    uc_key = "Update_Count"
-    step_key = "Step"
-    shared_state_key = "shared_optimizer_state"
-
-    optim_state = dict()
-    name_to_initializer_map = {n.name:n for n in model.graph.initializer}
-    initializers_names = name_to_initializer_map.keys()
-    for weight in initializers_names:
-        per_weight_state = dict()
-        weight_shape = name_to_initializer_map[weight].dims
-        for moment in moment_keys:
-            per_weight_state[moment] = np.full(weight_shape, 2.5, dtype=np.float32)
-        if isinstance(optimizer, optim.AdamConfig):
-            per_weight_state[uc_key] = np.full([1], 5, dtype=np.int64)
-        optim_state[weight] = copy.deepcopy(per_weight_state)
-    if isinstance(optimizer, optim.LambConfig):
-        step_val = np.full([1], 5, dtype=np.int64)
-        optim_state[shared_state_key] = {step_key : step_val}
-    return optim_state
-
-def get_optim_state_from_state_dict(state_dict, optimizer):
-    if not (isinstance(optimizer, optim.AdamConfig) or isinstance(optimizer, optim.LambConfig)):
-        return dict()
-    
-    moment_keys = ["Moment_1", "Moment_2"]
-    uc_key = "Update_Count"
-    step_key = "Step"
-    shared_state_key = "shared_optimizer_state"
-
-    optim_state = dict()
-    for param_name, v in state_dict.items():
-        for moment in moment_keys:
-            if param_name.startswith(moment):
-                fp32_name = param_name.split(moment + '_')[-1]
-                if fp32_name not in optim_state:
-                    optim_state[fp32_name] = dict()
-                optim_state[fp32_name].update({moment: v})
-                break
-        if param_name.startswith(uc_key):
-            fp32_name = param_name.split(uc_key + '_')[-1]
-            if fp32_name not in optim_state:
-                optim_state[fp32_name] = dict()
-            optim_state[fp32_name].update({uc_key: v})
-        elif param_name == step_key:
-            optim_state[shared_state_key] = {step_key: v}
-    return optim_state
-
 # EXPERIMENTAL HELPER FUNCTIONS
 
 def bert_model_description(dynamic_shape=True):
@@ -538,7 +486,7 @@ def testToyBertLoadOptimState(optimizer, mixedprecision_enabled):
     # Create ORTTrainer and save initial state in a dict
     model = load_bert_onnx_model()
     model_desc = bert_model_description()
-    dummy_init_state = generate_dummy_optim_state(model, optimizer)
+    dummy_init_state = _test_commons.generate_dummy_optim_state(model, optimizer)
     trainer = orttrainer.ORTTrainer(model, model_desc, optim_config, options=opts)
     checkpoint._experimental_load_optimizer_state(trainer, dummy_init_state)
     
@@ -554,7 +502,7 @@ def testToyBertLoadOptimState(optimizer, mixedprecision_enabled):
     _ = trainer.eval_step(input_ids, segment_ids, input_mask, masked_lm_labels, next_sentence_labels)
     
     actual_state = checkpoint.experimental_state_dict(trainer)
-    actual_optim_state = get_optim_state_from_state_dict(actual_state, optimizer)
+    actual_optim_state = _test_commons.get_optim_state_from_state_dict(actual_state, optimizer)
     _test_helpers.assert_optim_state(dummy_init_state, actual_optim_state)
 
 @pytest.mark.parametrize("model_params", [
